refactor(webpush): align tests with integration responsibilities
